refactor(dashboard): extract 5 shared page components + refactor 10 consumer pages
Extends the components/dashboard/shared/ module (previously holding
AnnouncementsPage + InvoicesPage) with five more parametrized page
components. Consolidates ~1,550 LOC of duplicated admin/staff page
bodies into thin shells that import the shared component and pass
basePath/apiEndpoint/staffId/allowX props.

New shared components (LOC):
  - DisputesPage.tsx       (468) — dispute list with filters + table
  - DisputeDetailPage.tsx  (648) — single dispute view + evidence submission
  - FeedbackPage.tsx       (629) — user feedback list
  - RefundsPage.tsx        (409) — refunds list + filters
  - SubscriptionsPage.tsx  (379) — subscriptions list + filters

Admin pages refactored to shells (before → after LOC):
  - disputes/page.tsx              335 →  16
  - disputes/[disputeId]/page.tsx  466 →  21
  - feedback/page.tsx              615 →  15
  - refunds/page.tsx               291 →  13
  - subscriptions/page.tsx         343 →  14

Staff pages refactored to shells:
  - disputes/page.tsx              454 →  15
  - disputes/[disputeId]/page.tsx  383 →  25
  - feedback/page.tsx              603 →  16
  - refunds/page.tsx               393 →  14
  - subscriptions/page.tsx         368 →  16

Props pattern (matches AnnouncementsPage / InvoicesPage):
  - `basePath`: for client-side links back to the dashboard root
  - `apiEndpoint`: source of list/detail data (admin vs staff endpoint)
  - `staffId?`: passed through where sub-links need it
  - `allowEvidenceSubmission?`: gates the admin-only actions in DisputeDetailPage
  - `queryKeyPrefix?`: so admin and staff can coexist in the same React
    Query cache without stepping on each other

NOT extracted (intentionally):
  - PaymentsPage — admin uses a simple table + filters, staff uses Tabs
    + Avatar rows + refund dialogs. Meaningfully different UX — forcing
    them into one parametrized component with many feature flags would
    create worse code than the duplication it eliminates. The real fix
    is a UX reconciliation (make one dashboard match the other), which
    is a product decision, not a cleanup.

Net: -1,548 LOC. 10 page files are now 5-25 lines each. All existing
functionality preserved — every query, filter, mutation, dialog, table
column, and empty state is intact via the shared components.
