+++ app/dashboard/admin/subscriptions/page.tsx
@@ -1,343 +1,14 @@
 "use client";
 
-import { Button } from "@/components/ui/button";
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { Input } from "@/components/ui/input";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import { Skeleton } from "@/components/ui/skeleton";
-import { useQuery } from "@tanstack/react-query";
-import { useState } from "react";
-import {
-  RefreshCw,
-  AlertCircle,
-  CheckCircle,
-  XCircle,
-  Search,
-} from "lucide-react";
-import type { SubscriptionListItem } from "@/types/subscriptions";
-
-async function fetchSubscriptions(
-  page: number,
-  limit: number,
-  status?: string,
-  search?: string,
-) {
-  const offset = (page - 1) * limit;
-  const params = new URLSearchParams({
-    limit: limit.toString(),
-    offset: offset.toString(),
-  });
-  if (status) params.set("status", status);
-  if (search) params.set("search", search);
-
-  const response = await fetch(`/api/admin/subscriptions?${params}`);
-  if (!response.ok) {
-    throw new Error("Failed to fetch subscriptions");
-  }
-  return response.json();
-}
+import { SubscriptionsPage } from "@/components/dashboard/shared/SubscriptionsPage";
 
 export default function AdminSubscriptionsPage() {
-  const [page, setPage] = useState(1);
-  const [status, setStatus] = useState<string | undefined>();
-  const [search, setSearch] = useState("");
-  const [searchInput, setSearchInput] = useState("");
-  const limit = 20;
-
-  const { data, isLoading, error } = useQuery({
-    queryKey: ["admin-subscriptions", page, status, search],
-    queryFn: () => fetchSubscriptions(page, limit, status, search),
-    staleTime: 60 * 1000,
-  });
-
-  const handleSearch = () => {
-    setSearch(searchInput);
-    setPage(1);
-  };
-
-  const getStatusBadge = (status: string) => {
-    switch (status) {
-      case "active":
-        return (
-          <span className="inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium bg-green-100 text-green-800">
-            <CheckCircle className="w-3 h-3 mr-1" />
-            Active
-          </span>
-        );
-      case "expiring_soon":
-        return (
-          <span className="inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium bg-yellow-100 text-yellow-800">
-            <AlertCircle className="w-3 h-3 mr-1" />
-            Expiring Soon
-          </span>
-        );
-      case "expired":
-        return (
-          <span className="inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium bg-gray-100 text-gray-800">
-            <XCircle className="w-3 h-3 mr-1" />
-            Expired
-          </span>
-        );
-      default:
-        return null;
-    }
-  };
-
-  if (error) {
-    return (
-      <div className="flex items-center justify-center h-full">
-        <Card className="w-full max-w-md">
-          <CardHeader>
-            <CardTitle className="text-red-600">Error</CardTitle>
-          </CardHeader>
-          <CardContent>
-            <p className="text-gray-700">
-              {error instanceof Error
-                ? error.message
-                : "Failed to load subscriptions"}
-            </p>
-          </CardContent>
-        </Card>
-      </div>
-    );
-  }
-
-  // Use stats from API (includes total counts, not just current page)
-  const activeCount =
-    (data?.stats?.activeCount || 0) + (data?.stats?.expiringCount || 0);
-  const expiringCount = data?.stats?.expiringCount || 0;
-
   return (
-    <div className="space-y-6">
-      <div>
-        <h1 className="text-3xl font-bold text-gray-900">Subscriptions</h1>
-        <p className="text-gray-600 mt-1">
-          Manage platform subscription appointments
-        </p>
-      </div>
-
-      {/* Summary Cards */}
-      <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
-        <Card>
-          <CardContent className="pt-4">
-            <div className="flex items-center justify-between">
-              <div>
-                <p className="text-sm text-gray-500">Active Subscriptions</p>
-                <p className="text-2xl font-bold text-green-700">
-                  {activeCount}
-                </p>
-              </div>
-              <CheckCircle className="w-8 h-8 text-green-500" />
-            </div>
-          </CardContent>
-        </Card>
-
-        <Card>
-          <CardContent className="pt-4">
-            <div className="flex items-center justify-between">
-              <div>
-                <p className="text-sm text-gray-500">Expiring This Week</p>
-                <p className="text-2xl font-bold text-yellow-700">
-                  {expiringCount}
-                </p>
-              </div>
-              <AlertCircle className="w-8 h-8 text-yellow-500" />
-            </div>
-          </CardContent>
-        </Card>
-
-        <Card>
-          <CardContent className="pt-4">
-            <div className="flex items-center justify-between">
-              <div>
-                <p className="text-sm text-gray-500">Total Subscriptions</p>
-                <p className="text-2xl font-bold text-gray-900">
-                  {data?.pagination?.total || 0}
-                </p>
-              </div>
-              <RefreshCw className="w-8 h-8 text-blue-500" />
-            </div>
-          </CardContent>
-        </Card>
-      </div>
-
-      {/* Filters */}
-      <Card>
-        <CardContent className="pt-4">
-          <div className="flex flex-wrap gap-4 items-center">
-            <div className="relative flex-1 min-w-[200px] max-w-md">
-              <Search className="absolute left-3 top-1/2 transform -translate-y-1/2 w-4 h-4 text-gray-400" />
-              <Input
-                placeholder="Search by name or email..."
-                value={searchInput}
-                onChange={(e) => setSearchInput(e.target.value)}
-                onKeyDown={(e) => e.key === "Enter" && handleSearch()}
-                className="pl-10"
-              />
-            </div>
-
-            <Select
-              value={status}
-              onValueChange={(value) => {
-                setStatus(value === "all" ? undefined : value);
-                setPage(1);
-              }}
-            >
-              <SelectTrigger className="w-40">
-                <SelectValue placeholder="Status" />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value="all">All Status</SelectItem>
-                <SelectItem value="active">Active</SelectItem>
-                <SelectItem value="expiring_soon">Expiring Soon</SelectItem>
-                <SelectItem value="expired">Expired</SelectItem>
-              </SelectContent>
-            </Select>
-
-            <Button onClick={handleSearch}>Search</Button>
-
-            {(search || status) && (
-              <Button
-                variant="ghost"
-                onClick={() => {
-                  setSearch("");
-                  setSearchInput("");
-                  setStatus(undefined);
-                  setPage(1);
-                }}
-              >
-                Clear Filters
-              </Button>
-            )}
-          </div>
-        </CardContent>
-      </Card>
-
-      <Card>
-        <CardHeader>
-          <CardTitle className="text-lg flex items-center gap-2">
-            <RefreshCw className="w-5 h-5 text-blue-600" />
-            Subscriptions
-          </CardTitle>
-        </CardHeader>
-        <CardContent>
-          {isLoading ? (
-            <div className="space-y-3">
-              {[1, 2, 3, 4, 5].map((i) => (
-                <Skeleton key={i} className="h-16 w-full" />
-              ))}
-            </div>
-          ) : data?.subscriptions?.length > 0 ? (
-            <>
-              <div className="overflow-x-auto">
-                <table className="w-full">
-                  <thead className="bg-gray-50 border-b">
-                    <tr>
-                      <th className="px-4 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                        Subscriber
-                      </th>
-                      <th className="px-4 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                        Consultant
-                      </th>
-                      <th className="px-4 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                        Amount
-                      </th>
-                      <th className="px-4 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                        Period
-                      </th>
-                      <th className="px-4 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                        Status
-                      </th>
-                    </tr>
-                  </thead>
-                  <tbody className="bg-white divide-y divide-gray-200">
-                    {data.subscriptions.map((sub: SubscriptionListItem) => (
-                      <tr key={sub.id} className="hover:bg-gray-50">
-                        <td className="px-4 py-3">
-                          <div>
-                            <p className="text-sm font-medium text-gray-900">
-                              {sub.userName}
-                            </p>
-                            <p className="text-xs text-gray-500">
-                              {sub.userEmail}
-                            </p>
-                          </div>
-                        </td>
-                        <td className="px-4 py-3 text-sm text-gray-700">
-                          {sub.consultantName || "-"}
-                        </td>
-                        <td className="px-4 py-3 text-sm font-semibold text-gray-900">
-                          {(sub.amount / 100).toLocaleString("en-IN", {
-                            style: "currency",
-                            currency: sub.currency,
-                          })}
-                        </td>
-                        <td className="px-4 py-3">
-                          <div className="text-sm text-gray-700">
-                            {sub.startDate
-                              ? new Date(sub.startDate).toLocaleDateString()
-                              : "-"}
-                            {" → "}
-                            {sub.endDate
-                              ? new Date(sub.endDate).toLocaleDateString()
-                              : "-"}
-                          </div>
-                        </td>
-                        <td className="px-4 py-3">
-                          {getStatusBadge(sub.status)}
-                        </td>
-                      </tr>
-                    ))}
-                  </tbody>
-                </table>
-              </div>
-
-              {/* Pagination */}
-              {data?.pagination && (
-                <div className="flex items-center justify-between pt-4">
-                  <div className="text-sm text-gray-500">
-                    Showing {(page - 1) * limit + 1} to{" "}
-                    {Math.min(page * limit, data.pagination.total)} of{" "}
-                    {data.pagination.total} subscriptions
-                  </div>
-                  <div className="flex gap-2">
-                    <Button
-                      variant="outline"
-                      size="sm"
-                      onClick={() => setPage((p) => Math.max(1, p - 1))}
-                      disabled={page === 1}
-                    >
-                      Previous
-                    </Button>
-                    <Button
-                      variant="outline"
-                      size="sm"
-                      onClick={() => setPage((p) => p + 1)}
-                      disabled={!data.pagination.hasMore}
-                    >
-                      Next
-                    </Button>
-                  </div>
-                </div>
-              )}
-            </>
-          ) : (
-            <div className="text-center py-12">
-              <p className="text-gray-500">
-                {search || status
-                  ? "No subscriptions match your filters"
-                  : "No subscriptions found"}
-              </p>
-            </div>
-          )}
-        </CardContent>
-      </Card>
-    </div>
+    <SubscriptionsPage
+      basePath="/dashboard/admin"
+      apiEndpoint="/api/admin/subscriptions"
+      title="Subscriptions"
+      description="Manage platform subscription appointments"
+    />
   );
 }
