+++ app/dashboard/staff/[staffId]/(features)/subscriptions/page.tsx
@@ -1,364 +1,18 @@
 "use client";
 
-import { useState, useEffect, useCallback } from "react";
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { Badge } from "@/components/ui/badge";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from "@/components/ui/table";
-import {
-  Search,
-  RefreshCw,
-  Loader2,
-  CheckCircle,
-  Clock,
-  XCircle,
-  AlertTriangle,
-  Users,
-} from "lucide-react";
-import { useToast } from "@/hooks/use-toast";
-import type {
-  SubscriptionListItem,
-  SubscriptionListResponse,
-} from "@/types/subscriptions";
-
-const getStatusColor = (status: string) => {
-  switch (status.toLowerCase()) {
-    case "active":
-      return "bg-green-100 text-green-700 dark:bg-green-900 dark:text-green-300";
-    case "expiring_soon":
-      return "bg-yellow-100 text-yellow-700 dark:bg-yellow-900 dark:text-yellow-300";
-    case "expired":
-      return "bg-red-100 text-red-700 dark:bg-red-900 dark:text-red-300";
-    default:
-      return "bg-zinc-100 text-zinc-700 dark:bg-zinc-800 dark:text-zinc-300";
-  }
-};
-
-const getStatusIcon = (status: string) => {
-  switch (status.toLowerCase()) {
-    case "active":
-      return <CheckCircle className="h-3 w-3" />;
-    case "expiring_soon":
-      return <AlertTriangle className="h-3 w-3" />;
-    case "expired":
-      return <XCircle className="h-3 w-3" />;
-    default:
-      return <Clock className="h-3 w-3" />;
-  }
-};
-
-const formatCurrency = (amount: number, currency: string = "INR") => {
-  return new Intl.NumberFormat("en-IN", {
-    style: "currency",
-    currency: currency,
-    maximumFractionDigits: 0,
-  }).format(amount / 100);
-};
-
-const formatDate = (dateString: string | null) => {
-  if (!dateString) return "-";
-  return new Date(dateString).toLocaleDateString("en-IN", {
-    day: "numeric",
-    month: "short",
-    year: "numeric",
-  });
-};
+import { useParams } from "next/navigation";
+import { SubscriptionsPage } from "@/components/dashboard/shared/SubscriptionsPage";
 
 export default function StaffSubscriptionsPage() {
-  const { toast } = useToast();
-  const [subscriptions, setSubscriptions] = useState<SubscriptionListItem[]>(
-    [],
-  );
-  const [loading, setLoading] = useState(true);
-  const [searchQuery, setSearchQuery] = useState("");
-  const [statusFilter, setStatusFilter] = useState("all");
-  const [page, setPage] = useState(1);
-  const [total, setTotal] = useState(0);
-  const [hasMore, setHasMore] = useState(false);
-  const [stats, setStats] = useState({
-    activeCount: 0,
-    expiringCount: 0,
-    expiredCount: 0,
-  });
-  const limit = 20;
-
-  const [debouncedSearch, setDebouncedSearch] = useState("");
-
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      setDebouncedSearch(searchQuery);
-      setPage(1);
-    }, 300);
-    return () => clearTimeout(timer);
-  }, [searchQuery]);
-
-  const fetchSubscriptions = useCallback(async () => {
-    setLoading(true);
-    try {
-      const params = new URLSearchParams();
-      params.set("limit", limit.toString());
-      params.set("offset", ((page - 1) * limit).toString());
-      if (debouncedSearch) params.set("search", debouncedSearch);
-      if (statusFilter !== "all") params.set("status", statusFilter);
-
-      const response = await fetch(`/api/admin/subscriptions?${params}`);
-      if (!response.ok) throw new Error("Failed to fetch subscriptions");
-
-      const data: SubscriptionListResponse = await response.json();
-      setSubscriptions(data.subscriptions);
-      setStats(data.stats);
-      setTotal(data.pagination.total);
-      setHasMore(data.pagination.hasMore);
-    } catch (error) {
-      console.error("Error fetching subscriptions:", error);
-      toast({
-        title: "Error",
-        description: "Failed to load subscriptions",
-        variant: "destructive",
-      });
-    } finally {
-      setLoading(false);
-    }
-  }, [page, debouncedSearch, statusFilter, toast]);
-
-  useEffect(() => {
-    fetchSubscriptions();
-  }, [fetchSubscriptions]);
-
-  const totalPages = Math.ceil(total / limit);
-
+  const params = useParams();
+  const staffId = params.staffId as string;
   return (
-    <div className="space-y-6">
-      {/* Header */}
-      <div className="flex items-center justify-between">
-        <div>
-          <h1 className="text-2xl font-bold text-zinc-900 dark:text-zinc-100">
-            Subscriptions
-          </h1>
-          <p className="text-zinc-500 dark:text-zinc-400">
-            View platform subscriptions (read-only)
-          </p>
-        </div>
-        <Button
-          variant="outline"
-          onClick={fetchSubscriptions}
-          disabled={loading}
-        >
-          <RefreshCw
-            className={`h-4 w-4 mr-2 ${loading ? "animate-spin" : ""}`}
-          />
-          Refresh
-        </Button>
-      </div>
-
-      {/* Stats */}
-      <div className="grid gap-4 md:grid-cols-4">
-        <Card>
-          <CardContent className="p-4 flex items-center gap-4">
-            <div className="p-2 rounded-lg bg-green-50 dark:bg-green-950">
-              <CheckCircle className="h-5 w-5 text-green-600" />
-            </div>
-            <div>
-              <p className="text-2xl font-bold">{stats.activeCount}</p>
-              <p className="text-sm text-zinc-500">Active</p>
-            </div>
-          </CardContent>
-        </Card>
-        <Card
-          className={
-            stats.expiringCount > 0
-              ? "border-yellow-200 bg-yellow-50 dark:bg-yellow-950/20"
-              : ""
-          }
-        >
-          <CardContent className="p-4 flex items-center gap-4">
-            <div className="p-2 rounded-lg bg-yellow-100 dark:bg-yellow-900">
-              <AlertTriangle className="h-5 w-5 text-yellow-600" />
-            </div>
-            <div>
-              <p className="text-2xl font-bold text-yellow-600">
-                {stats.expiringCount}
-              </p>
-              <p className="text-sm text-zinc-500">Expiring Soon</p>
-            </div>
-          </CardContent>
-        </Card>
-        <Card>
-          <CardContent className="p-4 flex items-center gap-4">
-            <div className="p-2 rounded-lg bg-red-50 dark:bg-red-950">
-              <XCircle className="h-5 w-5 text-red-600" />
-            </div>
-            <div>
-              <p className="text-2xl font-bold">{stats.expiredCount}</p>
-              <p className="text-sm text-zinc-500">Expired</p>
-            </div>
-          </CardContent>
-        </Card>
-        <Card className="bg-gradient-to-r from-purple-50 to-indigo-50 dark:from-purple-950/20 dark:to-indigo-950/20">
-          <CardContent className="p-4 flex items-center gap-4">
-            <div className="p-2 rounded-lg bg-purple-100 dark:bg-purple-900">
-              <Users className="h-5 w-5 text-purple-600" />
-            </div>
-            <div>
-              <p className="text-2xl font-bold text-purple-600">{total}</p>
-              <p className="text-sm text-zinc-500">Total Subscriptions</p>
-            </div>
-          </CardContent>
-        </Card>
-      </div>
-
-      {/* Filters */}
-      <Card>
-        <CardContent className="p-4">
-          <div className="flex flex-col sm:flex-row gap-4">
-            <div className="relative flex-1">
-              <Search className="absolute left-3 top-1/2 -translate-y-1/2 h-4 w-4 text-zinc-400" />
-              <Input
-                placeholder="Search by user name or email..."
-                className="pl-9"
-                value={searchQuery}
-                onChange={(e) => setSearchQuery(e.target.value)}
-              />
-            </div>
-            <Select
-              value={statusFilter}
-              onValueChange={(v) => {
-                setStatusFilter(v);
-                setPage(1);
-              }}
-            >
-              <SelectTrigger className="w-full sm:w-44">
-                <SelectValue placeholder="Status" />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value="all">All Status</SelectItem>
-                <SelectItem value="active">Active</SelectItem>
-                <SelectItem value="expiring_soon">Expiring Soon</SelectItem>
-                <SelectItem value="expired">Expired</SelectItem>
-              </SelectContent>
-            </Select>
-          </div>
-        </CardContent>
-      </Card>
-
-      {/* Subscriptions Table */}
-      <Card>
-        <CardHeader>
-          <CardTitle className="text-lg flex items-center gap-2">
-            <RefreshCw className="h-5 w-5 text-purple-600" />
-            Subscriptions ({total})
-          </CardTitle>
-        </CardHeader>
-        <CardContent className="p-0">
-          {loading ? (
-            <div className="flex items-center justify-center h-64">
-              <Loader2 className="h-8 w-8 animate-spin text-zinc-400" />
-            </div>
-          ) : subscriptions.length === 0 ? (
-            <div className="flex flex-col items-center justify-center h-64 text-zinc-500">
-              <RefreshCw className="h-12 w-12 mb-4 text-zinc-300" />
-              <p>No subscriptions found</p>
-            </div>
-          ) : (
-            <Table>
-              <TableHeader>
-                <TableRow>
-                  <TableHead>User</TableHead>
-                  <TableHead>Consultant</TableHead>
-                  <TableHead>Amount</TableHead>
-                  <TableHead>Period</TableHead>
-                  <TableHead>Status</TableHead>
-                  <TableHead>Gateway</TableHead>
-                </TableRow>
-              </TableHeader>
-              <TableBody>
-                {subscriptions.map((subscription) => (
-                  <TableRow key={subscription.id}>
-                    <TableCell>
-                      <div>
-                        <p className="font-medium">{subscription.userName}</p>
-                        <p className="text-xs text-zinc-400">
-                          {subscription.userEmail}
-                        </p>
-                      </div>
-                    </TableCell>
-                    <TableCell className="text-sm text-zinc-600 dark:text-zinc-400">
-                      {subscription.consultantName || "-"}
-                    </TableCell>
-                    <TableCell className="font-medium">
-                      {formatCurrency(
-                        subscription.amount,
-                        subscription.currency,
-                      )}
-                    </TableCell>
-                    <TableCell>
-                      <div className="text-sm">
-                        <p>{formatDate(subscription.startDate)}</p>
-                        <p className="text-zinc-400">
-                          to {formatDate(subscription.endDate)}
-                        </p>
-                      </div>
-                    </TableCell>
-                    <TableCell>
-                      <Badge
-                        className={`${getStatusColor(subscription.status)} gap-1`}
-                        variant="secondary"
-                      >
-                        {getStatusIcon(subscription.status)}
-                        {subscription.status.replace(/_/g, " ")}
-                      </Badge>
-                    </TableCell>
-                    <TableCell className="text-sm text-zinc-600 dark:text-zinc-400">
-                      {subscription.gateway}
-                    </TableCell>
-                  </TableRow>
-                ))}
-              </TableBody>
-            </Table>
-          )}
-        </CardContent>
-      </Card>
-
-      {/* Pagination */}
-      {totalPages > 1 && (
-        <div className="flex items-center justify-between">
-          <div className="text-sm text-zinc-500">
-            Showing {(page - 1) * limit + 1} to {Math.min(page * limit, total)}{" "}
-            of {total}
-          </div>
-          <div className="flex gap-2">
-            <Button
-              variant="outline"
-              disabled={page <= 1}
-              onClick={() => setPage(page - 1)}
-            >
-              Previous
-            </Button>
-            <Button
-              variant="outline"
-              disabled={!hasMore}
-              onClick={() => setPage(page + 1)}
-            >
-              Next
-            </Button>
-          </div>
-        </div>
-      )}
-    </div>
+    <SubscriptionsPage
+      basePath={`/dashboard/staff/${staffId}`}
+      staffId={staffId}
+      apiEndpoint="/api/admin/subscriptions"
+      title="Subscriptions"
+      description="View platform subscriptions (read-only)"
+    />
   );
 }
